refactor onadata.libs.utils.google to avoid pylint import issues with google-auth package
